+++ packages/reporter/src/components/detail/attachments/attachments.vue
@@ -127,6 +127,12 @@ const initList = (attachments) => {
         }
     }
 
+    // last group
+    if (group) {
+        sortGroupList(group);
+        list.push(group);
+    }
+
     return list;
 };