+++ packages/reporter/src/components/detail/attachments/audit.vue
@@ -3,6 +3,8 @@ import { watchEffect, shallowReactive } from 'vue';
 
 import Util from '../../../utils/util.js';
 
+import AttachmentHead from './attachment-head.vue';
+
 const props = defineProps({
     data: {
         type: Object,
@@ -37,12 +39,12 @@ watchEffect(() => {
     class="mcr-attachment-audit"
     open
   >
-    <summary class="mcr-attachment-head">
+    <AttachmentHead :retry="props.data.retry">
       <a
         :href="props.data.path"
         target="_blank"
       >{{ props.data.name }}</a>
-    </summary>
+    </AttachmentHead>
     <div class="mcr-attachment-body">
       <table>
         <tr class="mcr-head">