+++ packages/reporter/src/components/detail/attachments/network.vue
@@ -4,6 +4,7 @@ import { components } from 'vine-ui';
 
 import Util from '../../../utils/util.js';
 import generateWaterfallChart from '../../../modules/waterfall.js';
+import AttachmentHead from './attachment-head.vue';
 
 const { VuiFlex } = components;
 
@@ -96,12 +97,12 @@ watchEffect(() => {
     class="mcr-attachment-network"
     open
   >
-    <summary class="mcr-attachment-head">
+    <AttachmentHead :retry="props.data.retry">
       <a
         :href="props.data.path"
         target="_blank"
       >{{ props.data.name }}</a>
-    </summary>
+    </AttachmentHead>
     <VuiFlex
       class="mcr-attachment-body"
       padding="10px"