+++ packages/reporter/src/components/detail/attachments/video.vue
@@ -1,4 +1,5 @@
 <script setup>
+import AttachmentHead from './attachment-head.vue';
 const props = defineProps({
     data: {
         type: Object,
@@ -12,12 +13,12 @@ const props = defineProps({
     class="mcr-attachment-video"
     open
   >
-    <summary class="mcr-attachment-head">
+    <AttachmentHead :retry="props.data.retry">
       <a
         :href="props.data.path"
         target="_blank"
       >{{ props.data.name }}</a>
-    </summary>
+    </AttachmentHead>
     <div class="mcr-attachment-body">
       <video
         controls