+++ CHANGELOG.md
@@ -2,6 +2,7 @@
   - fixed coverage issue #29
   - updated search input #39  
   - fixed source type to use ext name
+  - fixed image diff not shown #47
 
 * 1.6.33
   - fixed tag row underline